Add auto-save functionality to settings.
Reworks the settings forms to not have a Save button, but instead do an
auto-save on each change.

This PR also reworks the settings so they're not rendered in a dialog,
but instead in a dedicated page, a la Discord.

The settings reducer has been moved to Immer and there's been some other
changes in here that I now forgot about :(