Write an implementation of onValidatedChange in the form hook.
This callback gets called whenever validations have completed
(successfully) for a change. There's some footgun potential here in that
we don't run validations for fields that were never "dirtied" (e.g. a
required text input that was never focused but is still empty will be
let through to this callback, although generally it shouldn't be a key
present in the object so maybe it's fine?). Anyway be careful with it.