Break the observer UI.
This is investigation into the flaky observer UI bug. It seems like the
flakes happen because CEF initializes early in the process, and then
times out (?) at some point an exits. Sometimes it exits early enough
that the SC:R will restart it when the game begins, other times it exits
*right as the game is beginning*, and the timing of things means it
doesn't get restarted.

In that flaky case, you can get it restarted afterwards by bringing up
the options menu, but that doesn't make the observer UI reload so =/

The "fix" here was an attempt to prevent it from loading early (after
trying to just ignore the call and return an error, which also didn't
work). With either of these "solutions", SC:R never seems to try to
re-launch CEF again, so things are just in a permanent state of broken.

WIP, obv.