Add db table for leagues, uuid v6 generator function, some basic model methods.
UUID v6 "solves" some issues with v4 as far as usage within a DB (v4 can
result in more expensive index usage/modification as it isn't sortable
based on time, whereas the most significant part of v6's bits are
directly based on time). This *should* be a drop-in replacement for any
of our UUID columns, although this commit doesn't go and replace the
default values anywhere yet.