Implement storing league user changes + leaderboards.
Top 10 is implemented here but I don't think I'm gonna keep it, as it
doesn't seem *that* useful, and if we were gonna have it we'd probably
want to display some minimal-info version (e.g. just rank, name,
points). I think in almost all cases you'd just want to go to the actual
leaderboard though, so this implementation complexity is unnecessary.